Move move_history attribute to Game class instead of Player
